alura-django-tickets/requirements.txt
Django==4.0.6
django-tempus-dominus==5.1.2.17
django-widget-tweaks==1.4.12
isort
black
